# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mercadeiro API
  description: >
    API do Mercadeiro — app de lista de compras domiciliar (Despensa → Carrinho →
    Finalizar Compra → Analítico). Documenta os endpoints da Fase 1 (API core).
    Regras de negócio referenciadas (RN-xx) e casos de uso (UC-xx) estão detalhados
    em `mercadeiro-regras-de-negocio.md`, na raiz do monorepo.


    **Autenticação:** todas as rotas exigem `Authorization: Bearer <token>`,
    exceto `/health`. Token único de aplicação (app single-user, sem contas —
    ver decisão em `mercadeiro-regras-de-negocio.md` §5.8), configurado em
    `API_BEARER_TOKEN` no `.env` do backend. Requisições sem token válido
    recebem `401 Unauthorized`.
  version: 0.1.0
servers:
  - url: http://localhost:3333
    description: Ambiente local de desenvolvimento
security:
  - bearerAuth: []
...
paths:
...
  /produtos:
    get:
      tags: [Produtos]
      summary: Lista produtos, com filtros opcionais
      operationId: listarProdutos
      parameters:
        - name: categoriaId
          in: query
          schema:
            type: string
        - name: busca
          in: query
          description: Busca por substring no nome (case-insensitive)
          schema:
            type: string
        - name: codigoBarras
          in: query
          schema:
            type: string
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Lista de produtos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProdutoComCategoria"
    post:
      tags: [Produtos]
      summary: Cria um produto ou reaproveita um existente com o mesmo nome
      description: >
        Regra 4.1: se já existir um Produto com o mesmo nome (case-insensitive),
        o registro existente é reaproveitado em vez de duplicado (status 200).
        Produto novo sem `categoriaId` válido cai na categoria "Outros".
      operationId: criarProduto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome:
                  type: string
                  minLength: 1
                categoriaId:
                  type: string
                  description: Se omitido ou inválido, cai no fallback "Outros"
                codigoBarras:
                  type: string
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Produto já existente reaproveitado (RN 4.1)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProdutoComCategoria"
        "201":
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProdutoComCategoria"
        "400":
          $ref: "#/components/responses/BadRequest"
  /produtos/{id}:
    parameters:
      - $ref: "#/components/parameters/ProdutoId"
    get:
      tags: [Produtos]
      summary: Busca um produto por id
      operationId: buscarProduto
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Produto encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProdutoComCategoria"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Produtos]
      summary: Atualiza nome, categoria e/ou código de barras de um produto
      operationId: atualizarProduto
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                  minLength: 1
                categoriaId:
                  type: string
                  description: Se inválido, cai no fallback "Outros"
                codigoBarras:
                  type: string
                  nullable: true
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Produto atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Produto"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Produtos]
      summary: Remove um produto sem uso na Despensa ou em Compras
      operationId: removerProduto
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "204":
          description: Produto removido
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Produto está em uso na Despensa ou em uma Compra
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /produtos/{id}/precos:
    parameters:
      - $ref: "#/components/parameters/ProdutoId"
    get:
      tags: [Produtos]
      summary: Histórico de preço do produto, agrupado por supermercado
      description: >
        Base do preço sugerido ao escanear (RN-03.7) e do ranking por supermercado
        (RN-07.3). `amostraPequena` é true com menos de 2 registros de PrecoHistorico
        (RN-07.4) — a informação é exibida mesmo assim, nunca ocultada.
      operationId: buscarPrecosDoProduto
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Agregados de preço do produto
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrecosProduto"
        "404":
          $ref: "#/components/responses/NotFound"
  /despensa:
    get:
      tags: [Despensa]
      summary: Lista os itens da Despensa agrupados por categoria
      description: >
        RN-02.2: grupos ordenados pela `ordem` da categoria; itens dentro do grupo
        ordenados por ordem de inserção. Categorias sem itens ativos não aparecem.
      operationId: listarDespensa
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Grupos de itens por categoria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GrupoDespensa"
    post:
      tags: [Despensa]
      summary: Adiciona um item à Despensa (cria produto se necessário)
      description: >
        Regra 4.1: se já existir um ItemDespensa para o produto (mesmo nome,
        case-insensitive), soma a quantidade em vez de duplicar a linha.
      operationId: adicionarItemDespensa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome:
                  type: string
                  minLength: 1
                quantidade:
                  type: integer
                  minimum: 1
                  default: 1
                categoriaId:
                  type: string
                codigoBarras:
                  type: string
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "201":
          description: Item criado ou quantidade somada ao item existente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ItemDespensa"
        "400":
          $ref: "#/components/responses/BadRequest"
  /despensa/{id}/quantidade:
    patch:
      tags: [Despensa]
      summary: Ajusta a quantidade de um item via incremento/decremento
      description: >
        RN-02.1: quantidade mínima é 1. Um delta que levaria a quantidade a 0 ou
        menos remove o item diretamente, sem confirmação.
      operationId: ajustarQuantidadeDespensa
      parameters:
        - $ref: "#/components/parameters/ItemDespensaId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [delta]
              properties:
                delta:
                  type: integer
                  description: Valor a somar à quantidade atual; não pode ser zero
                  example: -1
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Item removido ou item com quantidade atualizada
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ItemDespensa"
                  - $ref: "#/components/schemas/RemocaoConfirmada"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /despensa/{id}:
    delete:
      tags: [Despensa]
      summary: Remove um item da Despensa imediatamente
      operationId: removerItemDespensa
      parameters:
        - $ref: "#/components/parameters/ItemDespensaId"
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "204":
          description: Item removido
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    ProdutoId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    ItemDespensaId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Token único de aplicação (API_BEARER_TOKEN no .env do backend).
        Exigido em todas as rotas, exceto /health.
  responses:
...
    Unauthorized:
      description: Token de autenticação ausente ou inválido
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          description: >
            Detalhes adicionais, quando aplicável — formato varia por origem:
            array de issues do Zod em erros de validação, objeto de metadados
            do Prisma em conflitos de unicidade, ou ausente.
    RemocaoConfirmada:
      type: object
      required: [removido, id]
      properties:
        removido:
          type: boolean
          enum: [true]
        id:
          type: string
    Categoria:
      type: object
      required: [id, nome, ordem, createdAt, updatedAt]
      properties:
        id:
          type: string
        nome:
          type: string
        ordem:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Produto:
      type: object
      required: [id, nome, categoriaId, createdAt, updatedAt]
      properties:
        id:
          type: string
        nome:
          type: string
        codigoBarras:
          type: string
          nullable: true
        categoriaId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProdutoComCategoria:
      allOf:
        - $ref: "#/components/schemas/Produto"
        - type: object
          properties:
            categoria:
              $ref: "#/components/schemas/Categoria"
    Supermercado:
      type: object
      required: [id, nome, createdAt, updatedAt]
      properties:
        id:
          type: string
        nome:
          type: string
        endereco:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ItemDespensa:
      type: object
      required: [id, produtoId, quantidade, produto, createdAt, updatedAt]
      properties:
        id:
          type: string
        produtoId:
          type: string
        quantidade:
          type: integer
          minimum: 1
        produto:
          $ref: "#/components/schemas/ProdutoComCategoria"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    GrupoDespensa:
      type: object
      required: [categoria, itens]
      properties:
        categoria:
          $ref: "#/components/schemas/Categoria"
        itens:
          type: array
          items:
            $ref: "#/components/schemas/ItemDespensa"
...
    PrecoPorSupermercado:
      type: object
      required: [supermercado, precoMedio, quantidadeRegistros]
      properties:
        supermercado:
          $ref: "#/components/schemas/Supermercado"
        precoMedio:
          type: number
        quantidadeRegistros:
          type: integer
    PrecosProduto:
      type: object
      required: [ultimoPreco, menorPreco, precoMedio, amostraPequena, porSupermercado]
      properties:
        ultimoPreco:
          type: number
          nullable: true
          description: RN-07.1 — preço do PrecoHistorico mais recente, em qualquer supermercado.
        menorPreco:
          type: number
          nullable: true
        precoMedio:
          type: number
          nullable: true
        amostraPequena:
          type: boolean
          description: RN-07.4 — true com menos de 2 registros de PrecoHistorico; a informação parcial é exibida, não ocultada.
        porSupermercado:
          type: array
          items:
            $ref: "#/components/schemas/PrecoPorSupermercado"
...